continue family plugin

+++ src/libs/SaltyEmu.Communication/Protocol/RepositoryPacket/AsyncRpcRepository.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using ChickenAPI.Core.IPC;
+using ChickenAPI.Core.IPC.Protocol;
 using ChickenAPI.Data;
 
 namespace SaltyEmu.Communication.Protocol.RepositoryPacket
@@ -12,15 +13,19 @@ namespace SaltyEmu.Communication.Protocol.RepositoryPacket
         public AsyncRpcRepository(IAsyncRepository<T, TKey> repository, IIpcRequestHandler handler)
         {
             _repository = repository;
-            /*
-            handler.Register<RepositoryGetRequest<TKey>>(OnMessage);
-            handler.Register<RepositorySaveRequest<T>>(OnMessage);
-            handler.Register<RepositoryDeleteRequest<TKey>>(OnMessage);
-            */
+            // todo create delegates to remove the first parameter (which is supposed to be this)
+            handler.Register<RepositoryGetRequest<TKey>>(OnGet);
+            handler.Register<RepositorySaveRequest<T>>(OnSave);
+            handler.Register<RepositoryDeleteRequest<TKey>>(OnDelete);
         }
 
-        public async Task OnMessage(RepositoryGetRequest<TKey> request)
+        public async Task OnGet(IIpcRequest packet)
         {
+            if (!(packet is RepositoryGetRequest<TKey> request))
+            {
+                return;
+            }
+
             IEnumerable<T> tmp = await _repository.GetByIdsAsync(request.ObjectIds);
             await request.ReplyAsync(new RepositoryGetResponse<T>
             {
@@ -28,8 +33,12 @@ namespace SaltyEmu.Communication.Protocol.RepositoryPacket
             });
         }
 
-        public async Task OnMessage(RepositorySaveRequest<T> request)
+        public async Task OnSave(IIpcRequest packet)
         {
+            if (!(packet is RepositorySaveRequest<T> request))
+            {
+                return;
+            }
             await _repository.SaveAsync(request.Objects);
             await request.ReplyAsync(new RepositorySaveResponse<T>
             {
@@ -37,8 +46,12 @@ namespace SaltyEmu.Communication.Protocol.RepositoryPacket
             });
         }
 
-        public async Task OnMessage(RepositoryDeleteRequest<TKey> request)
+        public async Task OnDelete(IIpcRequest packet)
         {
+            if (!(packet is RepositoryDeleteRequest<TKey> request))
+            {
+                return;
+            }
             await _repository.DeleteByIdsAsync(request.ObjectIds);
         }
     }
